Moved copy schedule functions into a new file
